refactor: tidy up open, wsl and docker utils
